Added ability to change button colors via Core Graphics. Added related appearance properties.
